refactor hybrid mapper to use open_nudge_to_fine
